Add support for versioning
This MR will add the support for module upgrading based on versioning.
It adds two new functions, one to refresh the list of modules that need
update and another that will do the update.

More work needs to be done on the checheza-app and on the modules for
this to be 100% functional